lab5/requirements.txt
uvicorn
fastapi
httpx
pydantic
hazelcast-python-client
kafka-python
python-consul
